requirements.txt
# Core dependencies
faiss-cpu>=1.7.0       # Vector similarity search
pypdf>=3.0.0           # PDF text extraction
numpy>=1.24.0          # Array operations
requests>=2.28.0       # HTTP client for Ollama API

# For web article scraping
beautifulsoup4>=4.12.0 # HTML parsing

# For YouTube transcripts
youtube-transcript-api>=0.6.0  # YouTube transcript extraction

# Development
pytest>=7.0.0
